FIXED DIELECTRIC BUG!!!!!!!!!!!!!
Actually, it was not just limited to the dielectric, but all of my rays were only bouncing ONCE and not multiple times (limited by the recursion depth). The key was my switch statement was cutting the bounces short, and I was not passing the updated ray to the scattering functions (in subsequent calls to the scattering cuntions). WOOHOO!